fix provider zz_main.go generation; clean up files; test with make e2e
