+++ cluster/images/provider-upjet-alibabacloud/terraformrc.hcl
@@ -1,0 +1,9 @@
+provider_installation {
+  filesystem_mirror {
+    path    = "/terraform/provider-mirror"
+    include = ["*/*"]
+  }
+  direct {
+    exclude = ["*/*"]
+  }
+}
