+++ cmd/provider/alidns/zz_main.go
