+++ cmd/provider/cdn/zz_main.go
