+++ cmd/provider/cloudmonitorservice/zz_main.go
