+++ cmd/provider/ecs/zz_main.go
