+++ cmd/provider/tair/zz_main.go
