+++ hack/main.go.tmpl
@@ -40,7 +40,7 @@ import (
 
 func main() {
 	var (
-		app                     = kingpin.New(filepath.Base(os.Args[0]), "Terraform based Crossplane provider for AlibabaCloud {{ .Service }}").DefaultEnvars()
+		app                     = kingpin.New(filepath.Base(os.Args[0]), "Terraform based Crossplane provider for AlibabaCloud{{ if ne .Group "monolith" }} {{ .Group }}{{ end }}").DefaultEnvars()
 		debug                   = app.Flag("debug", "Run with debug logging.").Short('d').Bool()
 		syncPeriod              = app.Flag("sync", "Controller manager sync period such as 300ms, 1.5h, or 2h45m").Short('s').Default("1h").Duration()
 		pollInterval            = app.Flag("poll", "Poll interval controls how often an individual resource should be checked for drift.").Default("10m").Duration()
@@ -61,7 +61,7 @@ func main() {
 	kingpin.MustParse(app.Parse(os.Args[1:]))
 
 	zl := zap.New(zap.UseDevMode(*debug))
-	log := logging.NewLogrLogger(zl.WithName("provider-upjet-alibabacloud-{{ .Service }}"))
+	log := logging.NewLogrLogger(zl.WithName("provider-upjet-alibabacloud{{ if ne .Group "monolith" }}-{{ .Group }}{{ end }}"))
 	if *debug {
 		// The controller-runtime runs with a no-op logger by default. It is
 		// *very* verbose even at info level, so we only provide it a real
@@ -76,7 +76,7 @@ func main() {
 
 	mgr, err := ctrl.NewManager(cfg, ctrl.Options{
 		LeaderElection:   *leaderElection,
-		LeaderElectionID: "crossplane-leader-election-provider-upjet-alibabacloud-{{ .Service }}",
+		LeaderElectionID: "crossplane-leader-election-provider-upjet-alibabacloud{{ if ne .Group "monolith" }}-{{ .Group }}{{ end }}",
 		Cache: cache.Options{
 			SyncPeriod: syncPeriod,
 		},
@@ -147,6 +147,6 @@ func main() {
 		log.Info("Beta feature enabled", "flag", features.EnableBetaManagementPolicies)
 	}
 
-	kingpin.FatalIfError(controller.Setup_{{ .Service }}(mgr, o), "Cannot setup AlibabaCloud {{ .Service }} controllers")
+	kingpin.FatalIfError(controller.Setup_{{ .Group }}(mgr, o), "Cannot setup AlibabaCloud {{ .Group }} controllers")
 	kingpin.FatalIfError(mgr.Start(ctrl.SetupSignalHandler()), "Cannot start controller manager")
 }
