+++ internal/controller/zz_cdn_setup.go
@@ -14,7 +14,8 @@ import (
 	fctrigger "github.com/crossplane-contrib/provider-upjet-alibabacloud/internal/controller/cdn/fctrigger"
 )
 
-// Setup_cdn creates all CDN controllers for the CDN family provider.
+// Setup_cdn creates all controllers with the supplied logger and adds them to
+// the supplied manager.
 func Setup_cdn(mgr ctrl.Manager, o controller.Options) error {
 	for _, setup := range []func(ctrl.Manager, controller.Options) error{
 		domain.Setup,
