+++ internal/controller/zz_config_setup.go
@@ -9,15 +9,14 @@ import (
 
 	"github.com/crossplane/upjet/pkg/controller"
 
-	vpc "github.com/crossplane-contrib/provider-upjet-alibabacloud/internal/controller/vpc/vpc"
-	vswitch "github.com/crossplane-contrib/provider-upjet-alibabacloud/internal/controller/vpc/vswitch"
+	providerconfig "github.com/crossplane-contrib/provider-upjet-alibabacloud/internal/controller/providerconfig"
 )
 
-// Setup_vpc creates all VPC controllers for the VPC family provider.
-func Setup_vpc(mgr ctrl.Manager, o controller.Options) error {
+// Setup_config creates all controllers with the supplied logger and adds them to
+// the supplied manager.
+func Setup_config(mgr ctrl.Manager, o controller.Options) error {
 	for _, setup := range []func(ctrl.Manager, controller.Options) error{
-		vpc.Setup,
-		vswitch.Setup,
+		providerconfig.Setup,
 	} {
 		if err := setup(mgr, o); err != nil {
 			return err
