+++ internal/controller/zz_polardb_setup.go
@@ -22,7 +22,8 @@ import (
 	primaryendpoint "github.com/crossplane-contrib/provider-upjet-alibabacloud/internal/controller/polardb/primaryendpoint"
 )
 
-// Setup_polardb creates all POLARDB controllers for the POLARDB family provider.
+// Setup_polardb creates all controllers with the supplied logger and adds them to
+// the supplied manager.
 func Setup_polardb(mgr ctrl.Manager, o controller.Options) error {
 	for _, setup := range []func(ctrl.Manager, controller.Options) error{
 		account.Setup,
